CI: Add Tianocore edk2 as an external dependency
CI plugins (.pytool/Plugin/) and BaseTools are located in the
Tianocore edk2 repository. To re-use them, add the edk2 repository
as an external dependency.
Once the pip-requirements installed, the stuart_* set of commands
will download the required dependencies to locally run the CI. The
edk2 repository and its dependencies will be downloaded.

It is not currently possible to specify a branch name in the "version"
field. Thus, a fixed commit hash (current edk2 master) is written.
It will eventually be replaced by 'master'.

Signed-off-by: Pierre Gondois <Pierre.Gondois@arm.com>